Refactor: move create_contract in fuel_core_types::test_helpers
