+++ crates/types/src/lib.rs
@@ -76,3 +76,7 @@ pub mod fuel_vm {
         util,
     };
 }
+
+/// Helpers for creating transactions in tests
+#[cfg(feature = "test-helpers")]
+pub mod test_helpers;
